MIME-type and charset detection code refactoring:
Both are now able to use the actual content of the file
to improve the auto-detection.

Two simple heuristics were added (shebang line parsing
and Emacs mode parsing), more can be done later on that basis
(pluggable MIME-type detection scheme?).

The attachment, changeset and browser code that deal with this
has been adapted to call the common routines from mimeview/api.py
(`get_mimetype` function and `Mimeview.get_charset` method), 
giving to both of them a chunk of the file's content.

This closes #2105.
 

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@2653 af82e41b-90c4-0310-8c96-b1721e28e2e2