Autodetect big changesets and optimize their display.
When a changeset is deemed to be too big to be displayed in one page containing all the diffs, the actual diffs are simply not included in the view.

Instead, the modified file entries have a link pointing to a TracChangeset view that will contain only the changes affecting that file.

In order to detect big changesets, the following heuristic is used:
 1. The sum of the sizes in bytes of the modified files (old ones + new ones) should not exceed the `max_diff_bytes` parameter
 2. The total number of files in the changeset should not exceed the `max_diff_files` parameter.
 3. If there's only one file modified, rule 1. is ignored and the diff content is always shown.

The following defaults are used:
{{{
[changeset]
max_diff_bytes = 10000000
max_diff_files = 1000
}}}

This should close #2343 and #2591.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@2968 af82e41b-90c4-0310-8c96-b1721e28e2e2