Fixed the file viewer in the TracBrowser module.
There was some confusion between the requested path/rev and 
the last modified path/rev, which produced errors when viewing
a file which had just been copied. This fixes #2949.

Also, when no revision has been specified, the TracBrowser
don't artificially add the latest revision to its links.
This makes it possible to browse the repository to its 
latest state, unless explicitely making a revision ''sticky''
by specifying it in the ''View revision: '' form.
That fixes #987.

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@3099 af82e41b-90c4-0310-8c96-b1721e28e2e2