Some fixes and cleanups related to non-self contained repositories (#1830)
We only use the ''created path'' of some node if that created path stays 
within the scope of the repository.

Also, when generating changeset node change information,
we check whether the base path is within the scope or not,
and if we have the access right to it. 
In the negative, the base path is set to None, 
which means that edits or copies might then appear as added files.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@3174 af82e41b-90c4-0310-8c96-b1721e28e2e2