Robust coupling of Wiki lists and Wiki quotes
With [3316], new indent levels are only introduced when they need to be.

However, the implicit relationship that existed between the 
indentation level of the lists and the indentation level of the blocks
was lost.

We therefore keep track of the indentation level introduced by
either of them, referred to as "tabstops".

That way, the current indentation is remembered even when
empty lines are introduced for clarity, and the appropriate
indenting is used for quotes following lists or quotes:

{{{
     * first level
           * second level

           A second level quote
}}}

Tab characters are now converted before the line is parsed.

Fixes #124.

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@3318 af82e41b-90c4-0310-8c96-b1721e28e2e2