Make it possible to examine ticket description changes (#2945).
It is now possible to look at the history of the ticket description field and to visualize the actual changes.
It is also possible to view a ticket at a specified version, in read-only mode.

Next to the ''Description (last modified by <author>)'' title, there's now a (diff) link for showing the differences.
Each description change in the regular change history also features such a (diff) link.

From the diff page, one can go to a ''Ticket History'' view similar to the one of the Wiki page history, this one listing the changes for the ticket description field. From the history view, on can jump to a specific version of the ticket.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@4347 af82e41b-90c4-0310-8c96-b1721e28e2e2