Restore the usage of `IRequestFilter.post_process_request`, which was not enabled so far for Genshi templates.
Compared to the ClearSilver version, there's one additional parameter for the Genshi version: instead of `template` and `content_type`, it's now `template`, `data` and `content_type`. This makes it convenient to support both the 0.10 and 0.11 request filters, as we can distinguish them based on the arity of the `post_process_request` method.

The main usage for such a post-process filter is to change the data dictionary (aka template context) before the template is rendered.

The Timeline makes use of this to replace the `dateinfo` method, so that all the dateinfos become links to the timeline. Thanks to you-know-who for the clever idea ;)

Implements #975.

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@4649 af82e41b-90c4-0310-8c96-b1721e28e2e2