WorkFlow: enhanced the documentation and a few formatting cleanups.
In particular, I felt the need to clarify ''state'' vs. ''status''.
In many places, we were talking about ticket states referring to the ticket state as a whole, not only to its ''status'' field. The `ITicketActionController.get_all_states` method on the other hand was about getting the different values for the ''status'' fields, so I renamed the ITicketActionController method to `get_all_status`.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@5438 af82e41b-90c4-0310-8c96-b1721e28e2e2