Improved the layout of the ticket preview
So far, the preview was redirecting to the change preview block, which was at the bottom of the ticket ''Change History'' block. While this had the merit to show the change entry in the place it will take after the change is saved, this was a bit far away the top of the page, which also contains useful informations for the preview:
 - the occasional warnings in case of problems with the change
 - the new values for the ticket fields, among which the ticket description

Therefore, I moved that change preview block upward, just above the ticket box.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@5453 af82e41b-90c4-0310-8c96-b1721e28e2e2