Change Wiki navigation links from:
{{{
Start Page | Index by Title | Index by Date | Last Change
}}}

to:

{{{
Start Page | Index | History | Last Change
}}}

This enables direct access to both the revision log of a page and the diff corresponding to the last modification, see #5491.

A `trac-admin ... wiki upgrade` is needed, as the RecentChanges and TitleIndex pages now need to link to each other. The ''Index'' navigation link goes to the latter.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@5685 af82e41b-90c4-0310-8c96-b1721e28e2e2