On project creation, it is now possible to pass in an option for the `[inherit] file =` setting, and have the environment be a bit smarter on creation.
It will now load inherited settings when created, and not write defaults to project trac.ini except what is entered through the `initenv` console session or passed in as options to `Environment.create()`.

Command line option looks like this (can be appended to both variations of `initenv`):
{{{
trac-admin /path/to/new/env initenv --inherit=/path/to/global/trac.ini
}}}

Also includes a fix for not writing `[ticket-workflow]` when that section exists in the inherited trac.ini.

Closes #5651.


git-svn-id: http://svn.edgewall.org/repos/trac/trunk@6226 af82e41b-90c4-0310-8c96-b1721e28e2e2