Fix a bug in rendering CC: changes.
The comparison has to be done on the unobfuscated addresses, otherwise there's a possibility that no change will be detected in case there's a collision on the ''name@...'' part. This will have the undesirable effect that the change will be shown in the "changed from ''old'' to ''new''", with ''old'' and ''new'' lists  containing the full e-mail addresses.

Related to #153.

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@6242 af82e41b-90c4-0310-8c96-b1721e28e2e2